scan:  Add callbacks for scan results, add API to request scan.
Scan request method added to the UI object, but not implemented in
the android side.

Added callback to receive scan results to lfresource, but not yet
logic to actually send that to the LANforge manager.

Simplify the UI scan results slightly (save reference to ScanResults
object instead of getting it over and over).

Signed-off-by: Ben Greear <greearb@candelatech.com>