scan:  Implement scan callback to lfresource.
Add listener in the ResourceUtils the first time LANforge asks
for scan results.

Future callbacks will then send results to LANforge.

Ads some notes.  In particular, the current scan logic that populates
the UI is somewhat broken in that it does not use the callback handler,
and it asks way too often (radios take several seconds to do a scan, it
should ask maybe every 30 seconds or so).

Signed-off-by: Ben Greear <greearb@candelatech.com>